refactor(#176): unsprawl tools/recon/nmap.py into a package
Split the single nmap.py (~340 lines) into three focused files inside
a tools/recon/nmap/ package. Same public surface (nmap_scan +
port_scan), individually grep-able + testable internal layers.

Layout:

* tools/recon/nmap/__init__.py - re-exports nmap_scan + port_scan.
  Records the typed-shapes-live-in-models/ convention (NmapMode etc
  stay in models/network.py because they cross the agent boundary;
  tools/recon/nmap/* imports them one-way).
* tools/recon/nmap/flags.py - the config-tables layer. _MODE_FLAGS /
  _BANNER_INTENSITY / _SCRIPTS_EXPR / _SCAN_MODE_FLAGS dicts +
  _assemble_flags composer. No I/O; pure flag-composition.
* tools/recon/nmap/parser.py - _parse_xml using defusedxml. Defensive;
  routes banner strings through coerce_technologies on the way out.
* tools/recon/nmap/scanner.py - nmap_scan / port_scan / _evidence_filename
  / _resolve_evidence_path. The imperative orchestrator + the public
  function signatures.

The cybersquad-models discipline informed where to draw the line: the
LLM-facing typed shapes (NmapMode / NmapBanner / NmapScripts /
NmapService / NmapHostResult / NmapScanResult) stay in models/network.py
because the agent reads NmapScanResult from JSON across the boundary;
the implementation files (flags / parser / scanner) live in tools/ and
import the shapes one-way. No circular import surface.

tests/tools/recon/test_nmap.py - import paths repointed at the new
submodules (flags / parser / scanner); patches at the scanner module
boundary (tools.recon.nmap.scanner._require_binary etc) instead of the
top-level alias.

Test surface: 1744 still passing (unchanged - pure restructure, no
behaviour change). Ruff + mypy clean.

Next slice on this branch: rename tools/recon/probe.py -> httpx.py and
likely split into a similar package (flags / scanner / parser) once
verified the rename doesn't expose a stdlib-masking issue.
